# confluence/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Confluence loader",
    "version": "0.1.0"
  },
  "paths": {
    "/load-data-page-ids": {
      "post": {
        "summary": "Load data from Confluence filter by pageIds",
        "operationId": "load_data_load_data_page_ids_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestBodyWithPageIds"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Document"
                  },
                  "type": "array",
                  "title": "Response Load Data Load Data Page Ids Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/load-data-space-key": {
      "post": {
        "summary": "Load data from Confluence filter by space key",
        "operationId": "load_data_load_data_space_key_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestBodyWithSpaceKey"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Document"
                  },
                  "type": "array",
                  "title": "Response Load Data Load Data Space Key Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/load-data-label": {
      "post": {
        "summary": "Load data from Confluence filter by label",
        "operationId": "load_data_load_data_label_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestBodyWithLabel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Document"
                  },
                  "type": "array",
                  "title": "Response Load Data Load Data Label Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/load-data-cql": {
      "post": {
        "summary": "Load data from Confluence filter by CQL",
        "operationId": "load_data_load_data_cql_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestBodyWithCQL"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Document"
                  },
                  "type": "array",
                  "title": "Response Load Data Load Data Cql Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Document": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "doc_id": {
            "type": "string",
            "title": "Doc Id"
          },
          "extra_info": {
            "type": "object",
            "title": "Extra Info"
          }
        },
        "type": "object",
        "required": [
          "text",
          "doc_id",
          "extra_info"
        ],
        "title": "Document"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OauthScheme": {
        "properties": {
          "client_id": {
            "type": "string",
            "title": "Client Id"
          },
          "token": {
            "$ref": "#/components/schemas/TokenScheme"
          }
        },
        "type": "object",
        "required": [
          "client_id",
          "token"
        ],
        "title": "OauthScheme"
      },
      "RequestBodyWithCQL": {
        "properties": {
          "base_url": {
            "type": "string",
            "title": "Base Url",
            "default": "https://yoursite.atlassian.com/wiki"
          },
          "cql": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cql"
          },
          "start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "max_num_results": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Num Results"
          },
          "oauth2": {
            "$ref": "#/components/schemas/OauthScheme"
          }
        },
        "type": "object",
        "title": "RequestBodyWithCQL"
      },
      "RequestBodyWithLabel": {
        "properties": {
          "base_url": {
            "type": "string",
            "title": "Base Url",
            "default": "https://yoursite.atlassian.com/wiki"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "include_attachments": {
            "type": "boolean",
            "title": "Include Attachments",
            "default": true
          },
          "start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "max_num_results": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Num Results"
          },
          "oauth2": {
            "$ref": "#/components/schemas/OauthScheme"
          }
        },
        "type": "object",
        "title": "RequestBodyWithLabel"
      },
      "RequestBodyWithPageIds": {
        "properties": {
          "base_url": {
            "type": "string",
            "title": "Base Url",
            "default": "https://yoursite.atlassian.com/wiki"
          },
          "page_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Ids"
          },
          "include_attachments": {
            "type": "boolean",
            "title": "Include Attachments",
            "default": true
          },
          "include_children": {
            "type": "boolean",
            "title": "Include Children",
            "default": false
          },
          "start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "max_num_results": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Num Results"
          },
          "oauth2": {
            "$ref": "#/components/schemas/OauthScheme"
          }
        },
        "type": "object",
        "title": "RequestBodyWithPageIds"
      },
      "RequestBodyWithSpaceKey": {
        "properties": {
          "base_url": {
            "type": "string",
            "title": "Base Url",
            "default": "https://yoursite.atlassian.com/wiki"
          },
          "space_key": {
            "type": "string",
            "title": "Space Key"
          },
          "page_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Status"
          },
          "include_attachments": {
            "type": "boolean",
            "title": "Include Attachments",
            "default": true
          },
          "start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "max_num_results": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Num Results"
          },
          "oauth2": {
            "$ref": "#/components/schemas/OauthScheme"
          }
        },
        "type": "object",
        "title": "RequestBodyWithSpaceKey"
      },
      "TokenScheme": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "TokenScheme"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}